Convert StartActivity into a Fragment
This commit renames the StartActivity to StartFragment and updates the layout file and other related classes to reflect this change.
