+++ app/res/layout/start.xml
@@ -21,7 +21,7 @@
               android:layout_width="match_parent"
               android:layout_height="match_parent"
               android:orientation="vertical"
-              tools:context=".view.StartActivity">
+              tools:context=".view.StartFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/start_actionbar"
