+++ app/src/main/org/runnerup/view/ManageWorkoutsActivity.java
@@ -42,21 +42,11 @@ import android.widget.LinearLayout;
 import android.widget.RadioButton;
 import android.widget.TextView;
 import android.widget.Toast;
+
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.preference.PreferenceManager;
-import java.io.BufferedInputStream;
-import java.io.BufferedOutputStream;
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
+
 import org.runnerup.R;
 import org.runnerup.common.util.Constants;
 import org.runnerup.content.WorkoutFileProvider;
@@ -69,6 +59,19 @@ import org.runnerup.export.Synchronizer.Status;
 import org.runnerup.util.ViewUtil;
 import org.runnerup.workout.Workout;
 import org.runnerup.workout.WorkoutSerializer;
+
+import java.io.BufferedInputStream;
+import java.io.BufferedOutputStream;
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
 
 public class ManageWorkoutsActivity extends AppCompatActivity implements Constants {
 
@@ -259,7 +262,7 @@ public class ManageWorkoutsActivity extends AppCompatActivity implements Constan
 
     Intent intent =
         new Intent(this, MainLayout.class)
-            .putExtra("mode", StartActivity.TAB_ADVANCED)
+            .putExtra("mode", StartFragment.TAB_ADVANCED)
             .setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
     startActivity(intent);
     finish();
